+++ pkg/controller/common.go
@@ -108,7 +108,7 @@ func runGenerate(templatePath, dataPath, outputPath string, dryRun bool) error {
 
 	// Generate
 	conf.Logger.INFO(util.UR1, "Rendering template", nil)
-	bookUsecase := usecase.NewDefaultBookUsecase()
+	bookUsecase := usecase.NewBookUsecase(conf)
 	book, err := bookUsecase.Render(context.Background(), &gt, data)
 	if err != nil {
 		conf.Logger.ERROR(util.UR2, "Failed to render template", map[string]interface{}{"error": err.Error()})